# target/generated-sources/openapi/src/main/java/com/bank/debitcard/api/DebitCardsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bank.debitcard.api;

import com.bank.debitcard.model.DebitCardAccountLink;
import com.bank.debitcard.model.DebitCardCreateRequest;
import com.bank.debitcard.model.DebitCardResponse;
import com.bank.debitcard.model.DebitCardUpdateRequest;
import com.bank.debitcard.model.LinkAccountRequest;
import com.bank.debitcard.model.UnlinkAccountRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "debit-cards", description = "the debit-cards API")
public interface DebitCardsApi {

    /**
     * POST /debit-cards : Emitir una nueva tarjeta de débito
     * El número de tarjeta y el estado inicial los asigna el sistema (ACTIVE). No envíe cardNumber ni status en el body.
     *
     * @param debitCardCreateRequest  (required)
     * @return Tarjeta emitida (status code 201)
     *         or Solicitud inválida (status code 400)
     */
    @Operation(
        operationId = "createDebitCard",
        summary = "Emitir una nueva tarjeta de débito",
        description = "El número de tarjeta y el estado inicial los asigna el sistema (ACTIVE). No envíe cardNumber ni status en el body.",
        responses = {
            @ApiResponse(responseCode = "201", description = "Tarjeta emitida", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DebitCardResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/debit-cards",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<DebitCardResponse>> createDebitCard(
        @Parameter(name = "DebitCardCreateRequest", description = "", required = true) @Valid @RequestBody DebitCardCreateRequest debitCardCreateRequest
    );


    /**
     * DELETE /debit-cards/{id} : Baja lógica de la tarjeta
     * No elimina el registro; establece el estado en INACTIVE.
     *
     * @param id  (required)
     * @return Tarjeta dada de baja (INACTIVE) (status code 204)
     */
    @Operation(
        operationId = "deleteDebitCard",
        summary = "Baja lógica de la tarjeta",
        description = "No elimina el registro; establece el estado en INACTIVE.",
        responses = {
            @ApiResponse(responseCode = "204", description = "Tarjeta dada de baja (INACTIVE)")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/debit-cards/{id}"
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<Void>> deleteDebitCard(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    );


    /**
     * GET /debit-cards/{id}/accounts : Listar cuentas asociadas a la tarjeta
     *
     * @param id  (required)
     * @return Cuentas asociadas (status code 200)
     *         or Tarjeta no encontrada (status code 404)
     */
    @Operation(
        operationId = "getDebitCardAccounts",
        summary = "Listar cuentas asociadas a la tarjeta",
        responses = {
            @ApiResponse(responseCode = "200", description = "Cuentas asociadas", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = DebitCardAccountLink.class)))
            }),
            @ApiResponse(responseCode = "404", description = "Tarjeta no encontrada")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/debit-cards/{id}/accounts",
        produces = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<List<DebitCardAccountLink>>> getDebitCardAccounts(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    );


    /**
     * GET /debit-cards/{id} : Obtener tarjeta por ID
     *
     * @param id  (required)
     * @return Tarjeta encontrada (status code 200)
     *         or Tarjeta no encontrada (status code 404)
     */
    @Operation(
        operationId = "getDebitCardById",
        summary = "Obtener tarjeta por ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "Tarjeta encontrada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DebitCardResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Tarjeta no encontrada")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/debit-cards/{id}",
        produces = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<DebitCardResponse>> getDebitCardById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    );


    /**
     * GET /debit-cards : Listar tarjetas de débito activas
     * Devuelve únicamente tarjetas con estado ACTIVE.
     *
     * @return Lista de tarjetas (status code 200)
     */
    @Operation(
        operationId = "getDebitCards",
        summary = "Listar tarjetas de débito activas",
        description = "Devuelve únicamente tarjetas con estado ACTIVE.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Lista de tarjetas", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = DebitCardResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/debit-cards",
        produces = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<List<DebitCardResponse>>> getDebitCards(
        
    );


    /**
     * POST /debit-cards/{id}/link-account : Asociar una cuenta a la tarjeta
     *
     * @param id  (required)
     * @param linkAccountRequest  (required)
     * @return Asociación realizada (status code 200)
     *         or Tarjeta no encontrada (status code 404)
     *         or Regla de negocio no cumplida (status code 422)
     */
    @Operation(
        operationId = "linkAccountToDebitCard",
        summary = "Asociar una cuenta a la tarjeta",
        responses = {
            @ApiResponse(responseCode = "200", description = "Asociación realizada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DebitCardResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Tarjeta no encontrada"),
            @ApiResponse(responseCode = "422", description = "Regla de negocio no cumplida")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/debit-cards/{id}/link-account",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<DebitCardResponse>> linkAccountToDebitCard(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(name = "LinkAccountRequest", description = "", required = true) @Valid @RequestBody LinkAccountRequest linkAccountRequest
    );


    /**
     * POST /debit-cards/{id}/unlink-account : Desasociar una cuenta de la tarjeta
     *
     * @param id  (required)
     * @param unlinkAccountRequest  (required)
     * @return Desasociación realizada (status code 200)
     *         or Tarjeta no encontrada (status code 404)
     */
    @Operation(
        operationId = "unlinkAccountFromDebitCard",
        summary = "Desasociar una cuenta de la tarjeta",
        responses = {
            @ApiResponse(responseCode = "200", description = "Desasociación realizada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DebitCardResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Tarjeta no encontrada")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/debit-cards/{id}/unlink-account",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<DebitCardResponse>> unlinkAccountFromDebitCard(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(name = "UnlinkAccountRequest", description = "", required = true) @Valid @RequestBody UnlinkAccountRequest unlinkAccountRequest
    );


    /**
     * PUT /debit-cards/{id} : Actualizar tarjeta
     *
     * @param id  (required)
     * @param debitCardUpdateRequest  (required)
     * @return Tarjeta actualizada (status code 200)
     *         or Tarjeta no encontrada (status code 404)
     */
    @Operation(
        operationId = "updateDebitCard",
        summary = "Actualizar tarjeta",
        responses = {
            @ApiResponse(responseCode = "200", description = "Tarjeta actualizada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DebitCardResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Tarjeta no encontrada")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/debit-cards/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    io.reactivex.rxjava3.core.Single<ResponseEntity<DebitCardResponse>> updateDebitCard(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(name = "DebitCardUpdateRequest", description = "", required = true) @Valid @RequestBody DebitCardUpdateRequest debitCardUpdateRequest
    );

}