# target/generated-sources/openapi/src/main/java/com/bank/debitcard/model/DebitCardCreateRequest.java
package com.bank.debitcard.model;

import java.net.URI;
import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import org.openapitools.jackson.nullable.JsonNullable;
import java.time.OffsetDateTime;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;


import java.util.*;
import jakarta.annotation.Generated;

/**
 * DebitCardCreateRequest
 */

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.5.0")
public class DebitCardCreateRequest {

  private String customerId;

  public DebitCardCreateRequest() {
    super();
  }

  /**
   * Constructor with only required parameters
   */
  public DebitCardCreateRequest(String customerId) {
    this.customerId = customerId;
  }

  public DebitCardCreateRequest customerId(String customerId) {
    this.customerId = customerId;
    return this;
  }

  /**
   * ID del cliente titular
   * @return customerId
  */
  @NotNull 
  @Schema(name = "customerId", description = "ID del cliente titular", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("customerId")
  public String getCustomerId() {
    return customerId;
  }

  public void setCustomerId(String customerId) {
    this.customerId = customerId;
  }

  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    DebitCardCreateRequest debitCardCreateRequest = (DebitCardCreateRequest) o;
    return Objects.equals(this.customerId, debitCardCreateRequest.customerId);
  }

  @Override
  public int hashCode() {
    return Objects.hash(customerId);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class DebitCardCreateRequest {\n");
    sb.append("    customerId: ").append(toIndentedString(customerId)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}